Fix phantom note from extra space.
In cases where the available space for the keyboard is not evenly divisible by the number of white keys, some extra space exists to the right of the keyboard.  Without this fix, this space is playable as a phantom key that produces a low frequency.  You can reproduce this on the Qwerty Hancock demo page by clicking just to the right of the last white key.  This fix simply checks that the clicked element tag is 'li' prior to playing the note.