Make qwerty-hancock module exportable
Create checks for whether the environment was a browser or node. If the former, qwerty-hancock will know it is being used in the <script> context and carry on as usual. If the latter, qwerty-hancock will become exportable and can be required as a module. Add testing libraries, helpers, and specs to confirm this.