Enable qwerty-hancock to remove its events
An unexpected problem that occured was if the qwerty-hancock DOM element
was removed, its keyboard listeners still existed on the 'window'. A real
case example of this is if a user switched views, their computer keyboard
would still play sounds despite the keyboard UI not being visible.

This also prevents memory leaks by allowing users to remove events.